pyproject.toml
[tool.poetry]
name = "ska-tmc-common"
version = "0.20.0"
description = "SKA TMC common contains common classes required for Telescope Monitoring and Control(TMC)."
authors = ["Team Himalaya", "Team Sahyadri"]
license = "BSD-3-Clause"

packages = [
    { include = "ska_tmc_common", from = "src" },
]

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[[tool.poetry.source]]
name = "PyPI-public"
url = 'https://pypi.org/simple'

[tool.poetry.dependencies]
python = "^3.10"
numpy = [
    { version = ">=1.21.6", python = "3.10" },
    { version = ">=1.23.2", python = "3.11" },
    { version = ">=1.26.0", python = ">=3.12" },
]
pytango = "9.5.0"
jsonschema = "^4.17.3"
marshmallow = "^3.14.1"
fire = "^0.5.0"
ska-ser-logging = "^0.4.1"
ska-tango-base = "1.0.0"
ska-tango-testing = "0.6.0"
katpoint = "^1.0a2"
ska-telmodel = "^1.19.0"
mock = "^4.0.3"
typing_extensions = "*"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-mock = "^3.10.0"
pytest-xdist = "^2.4.0"
coverage = "^6.2"
pylint = "^3.1.0"
pylint2junit = "^1.0.1"
junitparser = "^2.2.0"
flake8-formatter-junit-xml = "^0.0.6"
flake8-rst-docstrings = "^0.3.0"
docstr-coverage = "^2.1.1"
darglint = "^1.8.1"
pylint-junit = "^0.3.2"
attrs = "^23.2.0"
referencing = "^0.35.0"
jsonschema-specifications = "^2023.12.1"
rpds-py = "^0.18.0"
mypy_extensions = "^1.0.0"
pathspec = "^0.12.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.10.1"
flake8 = "^7.0.0"

[tool.poetry.group.docs.dependencies]
docutils = "^0.18.1"
Sphinx = "^7.2.6"
ska-ser-sphinx-theme = "^0.1.2"
sphinx-autodoc-typehints = "^2.0.0"
sphinxcontrib-plantuml = "^0.22"
typing_extensions  = "^4.3.0"
sphinx-rtd-theme = "^1.2.2"
attrs = "*"
jsonschema = "*"
jsonschema-specifications = "*"
referencing = "*"
rpds-py = "*"
ska-control-model="*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
